Try hacking sys.path to make CLI work
