+++ tests/structured_app/conf/__init__.py
@@ -1,0 +1,40 @@
+from dataclasses import dataclass
+
+from hydra.core.config_store import ConfigStore
+from omegaconf import MISSING
+
+
+@dataclass
+class DBConfig:
+    driver: str = MISSING
+    host: str = "localhost"
+    port: int = MISSING
+
+
+@dataclass
+class MySQLConfig(DBConfig):
+    driver: str = "mysql"
+    port: int = 3306
+    user: str = MISSING
+    password: str = MISSING
+
+
+@dataclass
+class PostGreSQLConfig(DBConfig):
+    driver: str = "postgresql"
+    user: str = MISSING
+    port: int = 5432
+    password: str = MISSING
+    timeout: int = 10
+
+
+@dataclass
+class Config:
+    db: DBConfig = MISSING
+    debug: bool = False
+
+
+cs = ConfigStore.instance()
+cs.store(name="base_config", node=Config)
+cs.store(group="db", name="base_mysql", node=MySQLConfig)
+cs.store(group="db", name="base_postgresql", node=PostGreSQLConfig)
